+++ server/apps/rcrasite/models/__init__.py
@@ -1,5 +1,5 @@
-from .contact_models import Address, Contact, RcraPhone, RcraStates
-from .site_models import (
+from .contact import Address, Contact, RcraCountries, RcraPhone, RcraStates
+from .rcrasite import (
     RcraSite,
     RcraSiteType,
     Role,
