+++ server/apps/rcrasite/serializers/__init__.py
@@ -1,7 +1,3 @@
 from .address_serializer import AddressSerializer
 from .contact_serializer import ContactSerializer, RcraPhoneSerializer
-from .profile_serializer import (
-    RcrainfoSitePermissionsSerializer,
-    RcraSitePermissionSerializer,
-)
-from .site_serializer import RcraSiteSerializer
+from .site_serializer import RcraSiteSearchSerializer, RcraSiteSerializer
