+++ server/apps/rcrasite/services/__init__.py
@@ -1,2 +1,1 @@
-from .rcra_profile_services import RcraProfileService, get_or_create_rcra_profile
 from .rcra_site_services import RcraSiteService
